+++ attendance-api/src/routes/attendance/actions/invokeNextDay.ts
@@ -1,10 +1,11 @@
 import { Request, Response } from "express";
-import { BeneficiaryStatus } from "../../database/beneficiaryStatus";
-import { LogList } from "../../database/logList";
-import { logger } from "../../logger";
-import { BeneficiaryStatusModel } from "../../types/beneficiaryStatus";
-import { LogListModel } from "../../types/logList";
-import commonResponseError from "../../util/commonResponseError";
+import { BeneficiaryStatus } from "../../../database/beneficiaryStatus";
+import { LogList } from "../../../database/logList";
+import { OrderList } from "../../../database/orderList";
+import { logger } from "../../../logger";
+import { BeneficiaryStatusModel } from "../../../types/beneficiaryStatus";
+import { LogListModel } from "../../../types/logList";
+import commonResponseError from "../../../util/commonResponseError";
 
 export async function invokeNextDay(req: Request, res: Response) {
 	logger.info('invoking next day...')
@@ -22,8 +23,9 @@ export async function invokeNextDay(req: Request, res: Response) {
 			return logObject
 		}))
 
-		// reset status data
+		// reset status data and order list
 		await BeneficiaryStatus.updateAll({ status: 'absent' })
+		await OrderList.removeAll()
 		res.sendStatus(200)
 	} catch(err: any) {
 		commonResponseError(err, res)