+++ attendance-api/src/routes/attendance/orderList/removeFromList.ts
@@ -1,7 +1,8 @@
 import { Request, Response } from "express";
-import { OrderList } from "../../database/orderList";
-import { logger } from "../../logger";
-import commonResponseError from "../../util/commonResponseError";
+import { OrderList } from "../../../database/orderList";
+import { logger } from "../../../logger";
+import { emitMealServed } from "../../../socket";
+import commonResponseError from "../../../util/commonResponseError";
 
 export async function removeFromList(req: Request, res: Response) {
 	const { id } = req.params
@@ -9,6 +10,12 @@ export async function removeFromList(req: Request, res: Response) {
 
 	try {
 		await OrderList.updateById(parseInt(id), { active: false })
+
+		// notify meal served and beneficiary count
+		try {
+			await emitMealServed()
+		} catch {}
+		
 		logger.info('removed from order list')
 		res.sendStatus(200)
 	} catch(err: any) {