+++ attendance-api/src/routes/attendance/visitedList/removeById.ts
@@ -1,8 +1,8 @@
 import { Request, Response } from "express";
-import { VisitedUser } from "../../database/visitedUser";
-import { logger } from "../../logger";
-import { RemoveVisitedListResponse } from "../../types/visitedUser";
-import commonResponseError from "../../util/commonResponseError";
+import { VisitedUser } from "../../../database/visitedUser";
+import { logger } from "../../../logger";
+import { RemoveVisitedListResponse } from "../../../types/visitedUser";
+import commonResponseError from "../../../util/commonResponseError";
 
 export async function removeById(req: Request, res: Response) {
 	const { id } = req.params