+++ attendance-api/CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## v1.1.0 / 31/12/2022
+### **Added**
+- added "/attendance" before every route
+- attendance report implemented
+- socket implemented for live daily report (present, meal served and beneficiary count)
+
+### **Bug fixed**
+- reset order list when invoking next day
+
 ## v1.0.0 / 30/12/2022
 ### **Added**
 - basic CRUD functionalities implemented